fix: Don't fatal error when we can return an error (#36)
## Which problem is this PR solving?

Calling this code:
```go
	otelshutdown, err := otelconfig.ConfigureOpenTelemetry(
                 otelconfig.WithExporterProtocol(otelconfig.ProtocolHTTPJSON)))
```
Crashes with a fatal error, even though the function returns an error.

It would be nice to be able to catch this error and handle it in my app
appropriately.

## Short description of the changes

Return a wrapped error instead of crashing. Note that this also affects
other setup-type errors that are currently fatally crashing, and
required reworking some tests.

## How to verify that this has the expected result

Run the tests, or the code above; `err` should come back as non-nil and
it shouldn't crash.