feat: Add WithResourceOption for additional resource configuration (#48)
## Which problem is this PR solving?

We are unable to add resource detectors when configuring the launcher.

- Closes #12

## Short description of the changes

Implementation starts with what was written for #47 with smoke tests
added to verify the order of precedence for resources attributes from
environment over code.

[vera] upon further inspection, the original behavior of CODE attributes
winning over ENV attributes was valid:

> The SDK MUST extract information from the OTEL_RESOURCE_ATTRIBUTES
environment variable and
[merge](https://opentelemetry.io/docs/specs/otel/resource/sdk/#merge)
this, as the secondary resource, with any resource information provided
by the user, i.e. the user provided resource information has higher
priority.
— [OTel Resource SDK
spec](https://opentelemetry.io/docs/specs/otel/resource/sdk/#specifying-resource-information-via-an-environment-variable)

“user provided resource information” -- the user could mean the
application code writer OR the operator setting environment variables.
It’s ambiguous! And *not* settled in the spec:
https://github.com/open-telemetry/opentelemetry-specification/issues/1620

See note in that issue: 
>others in the Go SIG have interpreted this passage to require that
resource attributes from the environment must be subordinated to any
resource attributes provided by the developer-user

This is how the underlying Go SDK works, as designed.

The CODE>ENV decision was also made in the Ruby SDK:
https://github.com/robertlaurin/opentelemetry-ruby/blob/0c03e4aee09093ce0d80142e13a35326aad9f877/sdk/test/opentelemetry/sdk/configurator_test.rb#L35


Description from original PR:

Rather than adding a specific `WithDetectors` func to mirror the
functionality in the [resource
package](https://pkg.go.dev/go.opentelemetry.io/otel/sdk/resource#WithDetectors)
this adds a more generic version so that any resource option
configuration can be provided.

This has been done in a backwards compatible manner but I could see a
future where the `WithResourceAttributes` function is removed as it is
duplicate functionality now.

## How to verify that this has the expected result

- [x] Unit tests - logic's probably sound
- [x] Smoke tests - interaction with config from environment behaves
according to spec

---------

Co-authored-by: Martin Holman <me@martinholman.co.nz>
Co-authored-by: Vera Reynolds <verareynolds@honeycomb.io>